add refactor and new bill test
