+++ .travis.yml
@@ -3,6 +3,8 @@ python:
  - "3.6"
 # need to refactor to requirements/ setup
 install:
+ - pip install pipenv
  - pipenv install --dev
 script:
  - pylint nessie
+ - python -m unittest
