+++ nessie/models/bill.py
@@ -13,10 +13,19 @@
 }
 """
 
+# perhaps create a payment date?
 
-class _Bill():
-    def __init__(self):
-        pass
-    
-    
-    
+class Bill():
+    def __init__(self, _json):
+        self.id = _json["_id"]
+        self.status = _json["status"]
+        self.payee = _json["payee"]
+        self.nickname = _json["nickname"]
+        self.creation_date = _json["creation_date"]
+        self.payment_date = _json["payment_date"]
+        self.upcoming_payment_date = _json["upcoming_payment_date"]
+        self.recurring_date = _json["recurring_date"]
+        self.account_id = _json["account_id"]
+
+    def to_dict():
+        return vars(self)
